// packages/opml-poc/typechain-types/InferenceClaim.ts
export declare namespace InferenceClaim {
  export type ClaimStruct = {
    provider: AddressLike;
    modelHash: BytesLike;
    inputHash: BytesLike;
    outputHash: BytesLike;
    submittedAt: BigNumberish;
    challenger: AddressLike;
    status: BigNumberish;
  };

  export type ClaimStructOutput = [
    provider: string,
    modelHash: string,
    inputHash: string,
    outputHash: string,
    submittedAt: bigint,
    challenger: string,
    status: bigint
  ] & {
    provider: string;
    modelHash: string;
    inputHash: string;
    outputHash: string;
    submittedAt: bigint;
    challenger: string;
    status: bigint;
  };
}